LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/stylenonce-allowed.sub-expected.txt

This text should be green.

This text should also be green.

Style correctly allowed via a 'nonce-*' expression in 'style-src' should be applied to the page.


FAIL Should fire securitypolicyviolation assert_equals: expected "style-src-elem" but got "style-src"
PASS stylenonce-allowed
PASS stylenonce-allowed 1
